feat: 점수판 출력

+++ src/main/java/bowling/frame/result/Hit.java
@@ -39,6 +39,6 @@ public class Hit implements FrameResult {
 
 	@Override
 	public String toString() {
-		return String.valueOf(score);
+		return String.valueOf(score.toInt());
 	}
 }
